Version 2.0/requirements.txt
torch>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
streamlit>=1.28.0
Pillow>=10.0.0
